+++ src/factories/factoryBase.ts
@@ -33,6 +33,15 @@ export abstract class FactoryBase<T> {
     }
 
     /**
+     * List the types in the factory.
+     * @param name The name of the type to look for.
+     * @returns True if the type exists.
+     */
+    public types(): string[] {
+        return Object.keys(this.getInstance()._types);
+    }
+
+    /**
      * Create an instance of an object from the factory.
      * @param name The name of the type to create.
      * @param args Any parameters to pass to the constructor.